Async Statement API Development/src/main/java/com/example/statementgenerator/controller/StatementRequestController.java
package com.example.statementgenerator.controller;

import com.example.statementgenerator.model.StatementEntity;
import com.example.statementgenerator.service.StatementService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.concurrent.CompletableFuture;

@RestController
@RequestMapping("/api/v1/statements")
public class StatementRequestController {

    private final StatementService statementService;

    @Autowired
    public StatementRequestController(StatementService statementService) {
        this.statementService = statementService;
    }

    @PostMapping
    public CompletableFuture<ResponseEntity<String>> requestStatement(@RequestBody StatementEntity request) {
        return statementService.initiateStatementGeneration(request)
                .thenApply(requestId -> ResponseEntity.accepted().body("Request accepted with ID: " + requestId))
                .exceptionally(ex -> ResponseEntity.status(500).body("Error: " + ex.getMessage()));
    }
}
